TND场景 mask计算数据溢出 / kernel已知bug修复



# message auto-generated for no-merge-commit merge:
!1549 merge master into master

TND场景 mask计算数据溢出 / kernel已知bug修复

Created-by: liyangzhi
Commit-by: l00935273
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
修改TND + BSSmask场景下，mask拷贝过程数据计算溢出问题。将数据类型进行强制类型转换
修改kernel侧已知问题，数据块跨G轴时，部分数据没参与计算。

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->
DTS2025101311551、DTS2025101311582
## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-transformer-dev!1549

+++ attention/common/op_kernel/vector_common.h
@@ -791,7 +791,7 @@ struct MaskInfo {
 
 __aicore__ inline uint64_t ComputeAttenMaskOffsetNoCompress(MaskInfo &info, uint32_t s1StartIdx)
 {
-    uint64_t bOffset = info.batchIdx * info.batchOffset;
+    uint64_t bOffset = static_cast<uint64_t>(info.batchIdx) * static_cast<uint64_t>(info.batchOffset);
     uint64_t s1Offset = s1StartIdx % info.s1Size * info.attenMaskStride;
     uint64_t s2Offset = info.s2StartIdx;
     return bOffset + s1Offset + s2Offset;
